Gravity dataset without zero trade flows

An edited version of the full gravity dataset: The "square" gravity dataset for all possible pairs of Countries worldwide, 1948-2006, which is used in the article: Head, K., T. Mayer and J. Ries, 2010, "The erosion of colonial linkages after independence". Journal of International Economics, 81(1):1-14 (lead article). Dataset gravity_no_zeros corresponds to the dataset without zero trade flows, gravity_zeros, on the other hand, includes zero trade flows.

data("gravity_no_zeros")

Format

A data frame with 17088 observations on the following 10 variables.

iso_o

ISO-Code of country of origin

iso_d

ISO-Code of country of destination

distw

weighted distance

gdp_o

GDP of country of origin in million dollars

gdp_d

GDP of country of destination in million dollars

rta

regional trade agreement dummy

flow

trade flow in million dollars

contig

contiguity dummy

comlang_off

common official language dummy

comcur

common currency dummy

Source

http://econ.sciences-po.fr/node/131 https://sites.google.com/site/hiegravity/data-sources

References

Head, K. and Mayer, T. (2014). Chapter 3 - gravity equations: Workhorse,toolkit, and cookbook. In Gita Gopinath, E. H. and Rogoff, K., editors, Handbook of International Economics, volume 4 of Handbook of International Economics, pages 131-195. Elsevier. (Gravity Equations: Workhorse, Toolkit, and Cookbook)

Head, K., T. Mayer and J. Ries, 2010, "The erosion od colonial linkages after independence". Journal of International Economics, 81(1):1-14 (lead article).

Examples

  data(gravity_no_zeros)
  str(gravity_no_zeros)

#> tibble [17,088 × 10] (S3: tbl_df/tbl/data.frame)
#>  $ iso_o      : chr [1:17088] "AFG" "AFG" "AFG" "AFG" ...
#>  $ iso_d      : chr [1:17088] "ARG" "AUS" "AUT" "AZE" ...
#>  $ distw      : num [1:17088] 15341 11086 4567 1822 5332 ...
#>  $ gdp_o      : num [1:17088] 8399 8399 8399 8399 8399 ...
#>  $ gdp_d      : num [1:17088] 214058 768178 322444 20122 392001 ...
#>  $ rta        : num [1:17088] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ flow       : num [1:17088] 0.061 0.4055 0.1649 0.0016 4.1587 ...
#>  $ contig     : num [1:17088] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ comlang_off: num [1:17088] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ comcur     : num [1:17088] 0 0 0 0 0 0 0 0 0 0 ...
#>  - attr(*, "notes")= chr [1:7] "6" "data drawn from WDI for the year 2005 (earliest year available)." "the countries with no entry cost data are mainly small probably not in service trade data." "KOR designates RKO since 1949" ...