Negative Binomial Pseudo Maximum Likelihood (NBPML)

nbpml estimates gravity models in their multiplicative form via Negative Binomial Pseudo Maximum Likelihood.

nbpml(dependent_variable, distance, additional_regressors, robust = TRUE,
  data, ...)

Arguments

dependent_variable
(Type: character) name of the dependent variable. This variable is logged and then used as the dependent variable in the estimation.

distance
(Type: character) name of the distance variable that should be taken as the key independent variable in the estimation. The distance is logged automatically when the function is executed.

additional_regressors
(Type: character) names of the additional regressors to include in the model (e.g. a dummy variable to indicate contiguity). Unilateral metric variables such as GDPs can be added but those variables have to be logged first. Interaction terms can be added.

Write this argument as c(contiguity, common currency, ...). By default this is set to NULL.

robust
(Type: logical) whether robust fitting should be used. By default this is set to FALSE.

data
(Type: data.frame) the dataset to be used.

...
Additional arguments to be passed to the function.

Value

The function returns the summary of the estimated gravity model similar to a glm-object.

Details

nbpml is an estimation method for gravity models belonging to generalized linear models. It is estimated via glm.nb using the negative binomial distribution and a log-link.

For similar functions, utilizing the multiplicative form via the log-link, but different distributions, see nbpml, gpml, and nls.

gpml estimation can be used for both, cross-sectional as well as panel data, but its up to the user to ensure that the functions can be applied to panel data.

Depending on the panel dataset and the variables - specifically the type of fixed effects - included in the model, it may easily occur that the model is not computable.

Also, note that by including bilateral fixed effects such as country-pair effects, the coefficients of time-invariant observables such as distance can no longer be estimated.

Depending on the specific model, the code of the respective function may has to be changed in order to exclude the distance variable from the estimation.

At the very least, the user should take special care with respect to the meaning of the estimated coefficients and variances as well as the decision about which effects to include in the estimation. When using panel data, the parameter and variance estimation of the models may have to be changed accordingly.

For a comprehensive overview of gravity models for panel data see Egger and Pfaffermayr (2003) , Gómez-Herrera (2013) and Head et al. (2010) .

References

For more information on gravity models, theoretical foundations and estimation methods in general see

Anderson JE (1979). “A Theoretical Foundation for the Gravity Equation.” The American Economic Review, 69(1), 106--116. ISSN 00028282.

Anderson JE, van Wincoop E (2001). “Gravity with Gravitas: A Solution to the Border Puzzle.” Technical Report 8079, National Bureau of Economic Research. doi: 10.3386/w8079 .

Anderson JE (2010). “The Gravity Model.” Technical Report 16576, National Bureau of Economic Research. doi: 10.3386/w16576 .

Baier SL, Bergstrand JH (2009). “Bonus vetus OLS: A simple method for approximating international trade-cost effects using the gravity equation.” Journal of International Economics, 77(1), 77 - 85. ISSN 0022-1996, doi: 10.1016/j.jinteco.2008.10.004 .

Baier SL, Bergstrand JH (2010). “The Gravity Model in International Trade: Advances and Applications.” In van Bergeijk PAG, Brakman S (eds.), chapter 4. Cambridge University Press. doi: 10.1017/CBO9780511762109 .

Feenstra RC (2002). “Border effects and the gravity equation: consistent methods for estimation.” Scottish Journal of Political Economy, 49(5), 491--506.

Head K, Mayer T, Ries J (2010). “The erosion of colonial trade linkages after independence.” Journal of International Economics, 81(1), 1 - 14. ISSN 0022-1996, doi: 10.1016/j.jinteco.2010.01.002 .

Head K, Mayer T (2014). “Chapter 3 - Gravity Equations: Workhorse,Toolkit, and Cookbook.” In Gopinath G, Helpman E, Rogoff K (eds.), Handbook of International Economics, volume 4 series Handbook of International Economics, 131 - 195. Elsevier. doi: 10.1016/B978-0-444-54314-1.00003-3 .

Silva JMCS, Tenreyro S (2006). “The Log of Gravity.” The Review of Economics and Statistics, 88(4), 641-658. doi: 10.1162/rest.88.4.641 .

and the citations therein.

See Gravity Equations: Workhorse, Toolkit, and Cookbook for gravity datasets and Stata code for estimating gravity models.

For estimating gravity equations using panel data see

Egger P, Pfaffermayr M (2003). “The proper panel econometric specification of the gravity equation: A three-way model with bilateral interaction effects.” Empirical Economics, 28(3), 571--580. ISSN 1435-8921, doi: 10.1007/s001810200146 .

Gómez-Herrera E (2013). “Comparing alternative methods to estimate gravity models of bilateral trade.” Empirical Economics, 44(3), 1087--1111. ISSN 1435-8921, doi: 10.1007/s00181-012-0576-2 .

and the references therein.

See also

Examples

# Example for CRAN checks:
# Executable in < 5 sec
library(dplyr)
data("gravity_no_zeros")

# Choose 5 countries for testing
countries_chosen <- c("AUS", "CHN", "GBR", "BRA", "CAN")
grav_small <- filter(gravity_no_zeros, iso_o %in% countries_chosen)

fit <- nbpml(
  dependent_variable = "flow",
  distance = "distw",
  additional_regressors = c("rta", "iso_o", "iso_d"),
  robust = FALSE,
  data = grav_small
)

#> Warning: non-integer x = 16.451930

#> Warning: non-integer x = 4.503260

#> Warning: non-integer x = 0.467355

#> Warning: non-integer x = 142.120000

#> Warning: non-integer x = 0.868430

#> Warning: non-integer x = 0.578746

#> Warning: non-integer x = 96.275500

#> Warning: non-integer x = 8.415400

#> Warning: non-integer x = 0.369980

#> Warning: non-integer x = 1574.170000

#> Warning: non-integer x = 1.420034

#> Warning: non-integer x = 1.518352

#> Warning: non-integer x = 183.113000

#> Warning: non-integer x = 55.232600

#> Warning: non-integer x = 2.125680

#> Warning: non-integer x = 2.748000

#> Warning: non-integer x = 0.113788

#> Warning: non-integer x = 0.576175

#> Warning: non-integer x = 811.544000

#> Warning: non-integer x = 0.006378

#> Warning: non-integer x = 4.833477

#> Warning: non-integer x = 0.092658

#> Warning: non-integer x = 1549.610000

#> Warning: non-integer x = 317.366500

#> Warning: non-integer x = 202.811000

#> Warning: non-integer x = 19194.500000

#> Warning: non-integer x = 2.429410

#> Warning: non-integer x = 2.582130

#> Warning: non-integer x = 1.966100

#> Warning: non-integer x = 28.104010

#> Warning: non-integer x = 0.018180

#> Warning: non-integer x = 0.015766

#> Warning: non-integer x = 12.656930

#> Warning: non-integer x = 79.974180

#> Warning: non-integer x = 1349.040000

#> Warning: non-integer x = 1.391920

#> Warning: non-integer x = 0.812202

#> Warning: non-integer x = 168.267000

#> Warning: non-integer x = 6.172280

#> Warning: non-integer x = 84.557600

#> Warning: non-integer x = 6.984648

#> Warning: non-integer x = 345.982000

#> Warning: non-integer x = 0.776325

#> Warning: non-integer x = 1042.040000

#> Warning: non-integer x = 7.645143

#> Warning: non-integer x = 2.846426

#> Warning: non-integer x = 972.541900

#> Warning: non-integer x = 371.136000

#> Warning: non-integer x = 1294.090000

#> Warning: non-integer x = 0.995597

#> Warning: non-integer x = 6717.535000

#> Warning: non-integer x = 5.558564

#> Warning: non-integer x = 77.438400

#> Warning: non-integer x = 14.576210

#> Warning: non-integer x = 0.185703

#> Warning: non-integer x = 0.674837

#> Warning: non-integer x = 0.790087

#> Warning: non-integer x = 36.620650

#> Warning: non-integer x = 0.126019

#> Warning: non-integer x = 18.973100

#> Warning: non-integer x = 1.617040

#> Warning: non-integer x = 2571.195000

#> Warning: non-integer x = 6.000038

#> Warning: non-integer x = 12.440890

#> Warning: non-integer x = 0.407133

#> Warning: non-integer x = 19.437000

#> Warning: non-integer x = 3660.327000

#> Warning: non-integer x = 7412.760000

#> Warning: non-integer x = 154.882000

#> Warning: non-integer x = 341.696300

#> Warning: non-integer x = 12.730410

#> Warning: non-integer x = 1571.540000

#> Warning: non-integer x = 26.880810

#> Warning: non-integer x = 88.165600

#> Warning: non-integer x = 27775.800000

#> Warning: non-integer x = 5.675410

#> Warning: non-integer x = 47.461300

#> Warning: non-integer x = 4.178110

#> Warning: non-integer x = 35.485010

#> Warning: non-integer x = 22.425920

#> Warning: non-integer x = 0.176210

#> Warning: non-integer x = 20.580700

#> Warning: non-integer x = 20.545140

#> Warning: non-integer x = 0.576896

#> Warning: non-integer x = 8.702630

#> Warning: non-integer x = 1.875720

#> Warning: non-integer x = 139.529500

#> Warning: non-integer x = 0.093585

#> Warning: non-integer x = 3.574230

#> Warning: non-integer x = 9.257358

#> Warning: non-integer x = 73.187100

#> Warning: non-integer x = 28.200810

#> Warning: non-integer x = 0.384957

#> Warning: non-integer x = 5.695540

#> Warning: non-integer x = 18.961100

#> Warning: non-integer x = 986.199000

#> Warning: non-integer x = 1.775570

#> Warning: non-integer x = 11.224800

#> Warning: non-integer x = 14.151900

#> Warning: non-integer x = 261.207100

#> Warning: non-integer x = 4.025500

#> Warning: non-integer x = 99.439560

#> Warning: non-integer x = 0.506128

#> Warning: non-integer x = 2425.620000

#> Warning: non-integer x = 3.323672

#> Warning: non-integer x = 4.902502

#> Warning: non-integer x = 37.095740

#> Warning: non-integer x = 1.833780

#> Warning: non-integer x = 2368.520000

#> Warning: non-integer x = 107.293670

#> Warning: non-integer x = 19.570500

#> Warning: non-integer x = 7379.130000

#> Warning: non-integer x = 257.663000

#> Warning: non-integer x = 6.461697

#> Warning: non-integer x = 80.163000

#> Warning: non-integer x = 851.290000

#> Warning: non-integer x = 1.062173

#> Warning: non-integer x = 1228.051000

#> Warning: non-integer x = 74.575800

#> Warning: non-integer x = 52.123720

#> Warning: non-integer x = 5.985090

#> Warning: non-integer x = 90.930500

#> Warning: non-integer x = 556.097300

#> Warning: non-integer x = 0.109962

#> Warning: non-integer x = 235.571600

#> Warning: non-integer x = 10.540871

#> Warning: non-integer x = 3805.593000

#> Warning: non-integer x = 65.263660

#> Warning: non-integer x = 3.768880

#> Warning: non-integer x = 14.417700

#> Warning: non-integer x = 1.335180

#> Warning: non-integer x = 2.019380

#> Warning: non-integer x = 14.269000

#> Warning: non-integer x = 11.788500

#> Warning: non-integer x = 393.193000

#> Warning: non-integer x = 0.271145

#> Warning: non-integer x = 4.964360

#> Warning: non-integer x = 4.458380

#> Warning: non-integer x = 3.133702

#> Warning: non-integer x = 3.045075

#> Warning: non-integer x = 3563.296000

#> Warning: non-integer x = 0.301900

#> Warning: non-integer x = 0.604823

#> Warning: non-integer x = 10.464600

#> Warning: non-integer x = 18.921650

#> Warning: non-integer x = 3.483480

#> Warning: non-integer x = 365.747000

#> Warning: non-integer x = 26.987400

#> Warning: non-integer x = 4.431482

#> Warning: non-integer x = 67.300800

#> Warning: non-integer x = 15.129100

#> Warning: non-integer x = 8562.900000

#> Warning: non-integer x = 0.665409

#> Warning: non-integer x = 0.522070

#> Warning: non-integer x = 20.186320

#> Warning: non-integer x = 1258.150000

#> Warning: non-integer x = 55.824500

#> Warning: non-integer x = 27.978940

#> Warning: non-integer x = 123.673000

#> Warning: non-integer x = 0.532371

#> Warning: non-integer x = 1918.114000

#> Warning: non-integer x = 1.072401

#> Warning: non-integer x = 28.107100

#> Warning: non-integer x = 10.667800

#> Warning: non-integer x = 3.235617

#> Warning: non-integer x = 919.683600

#> Warning: non-integer x = 28.409700

#> Warning: non-integer x = 12885.180000

#> Warning: non-integer x = 17.354810

#> Warning: non-integer x = 6.199501

#> Warning: non-integer x = 681.470000

#> Warning: non-integer x = 161.198000

#> Warning: non-integer x = 85.288000

#> Warning: non-integer x = 0.254847

#> Warning: non-integer x = 3293.829000

#> Warning: non-integer x = 42.924100

#> Warning: non-integer x = 5.002680

#> Warning: non-integer x = 302.594600

#> Warning: non-integer x = 461.896000

#> Warning: non-integer x = 15.682700

#> Warning: non-integer x = 209.922000

#> Warning: non-integer x = 4.197600

#> Warning: non-integer x = 763.344000

#> Warning: non-integer x = 0.031215

#> Warning: non-integer x = 4.051432

#> Warning: non-integer x = 4.699950

#> Warning: non-integer x = 3308.870000

#> Warning: non-integer x = 930.444900

#> Warning: non-integer x = 4285.578000

#> Warning: non-integer x = 12907.200000

#> Warning: non-integer x = 82.781300

#> Warning: non-integer x = 114.893000

#> Warning: non-integer x = 37.504610

#> Warning: non-integer x = 2357.421000

#> Warning: non-integer x = 2.668061

#> Warning: non-integer x = 30.815400

#> Warning: non-integer x = 470.551000

#> Warning: non-integer x = 75.334900

#> Warning: non-integer x = 6525.050000

#> Warning: non-integer x = 8.304318

#> Warning: non-integer x = 1.943524

#> Warning: non-integer x = 286.354000

#> Warning: non-integer x = 400.541000

#> Warning: non-integer x = 517.854700

#> Warning: non-integer x = 960.658600

#> Warning: non-integer x = 1484.032000

#> Warning: non-integer x = 8.096297

#> Warning: non-integer x = 2673.040000

#> Warning: non-integer x = 55.787600

#> Warning: non-integer x = 59.963100

#> Warning: non-integer x = 519.259000

#> Warning: non-integer x = 0.885386

#> Warning: non-integer x = 3230.790000

#> Warning: non-integer x = 29.634900

#> Warning: non-integer x = 3566.600000

#> Warning: non-integer x = 52.267200

#> Warning: non-integer x = 245.225200

#> Warning: non-integer x = 32.461110

#> Warning: non-integer x = 46.492200

#> Warning: non-integer x = 4.162532

#> Warning: non-integer x = 11.610000

#> Warning: non-integer x = 271.506400

#> Warning: non-integer x = 5.181330

#> Warning: non-integer x = 351.284000

#> Warning: non-integer x = 22.092200

#> Warning: non-integer x = 1132.648000

#> Warning: non-integer x = 154.738100

#> Warning: non-integer x = 144.558000

#> Warning: non-integer x = 72.986800

#> Warning: non-integer x = 137.863000

#> Warning: non-integer x = 545.775000

#> Warning: non-integer x = 1030.244600

#> Warning: non-integer x = 388.452900

#> Warning: non-integer x = 1724.701000

#> Warning: non-integer x = 6.104020

#> Warning: non-integer x = 4327.680000

#> Warning: non-integer x = 199.139600

#> Warning: non-integer x = 124.787300

#> Warning: non-integer x = 5048.370000

#> Warning: non-integer x = 45.810710

#> Warning: non-integer x = 52.419620

#> Warning: non-integer x = 7.370610

#> Warning: non-integer x = 1.850508

#> Warning: non-integer x = 0.097701

#> Warning: non-integer x = 1.935970

#> Warning: non-integer x = 2423.920000

#> Warning: non-integer x = 0.129183

#> Warning: non-integer x = 195.950000

#> Warning: non-integer x = 12.571800

#> Warning: non-integer x = 224.715700

#> Warning: non-integer x = 4.441410

#> Warning: non-integer x = 83.538730

#> Warning: non-integer x = 0.059005

#> Warning: non-integer x = 45.063480

#> Warning: non-integer x = 48.473480

#> Warning: non-integer x = 7.687580

#> Warning: non-integer x = 430.600500

#> Warning: non-integer x = 57.427600

#> Warning: non-integer x = 7.551434

#> Warning: non-integer x = 2.757590

#> Warning: non-integer x = 6113.620000

#> Warning: non-integer x = 39.637900

#> Warning: non-integer x = 14.006800

#> Warning: non-integer x = 1.912427

#> Warning: non-integer x = 38.488890

#> Warning: non-integer x = 80.792470

#> Warning: non-integer x = 9.156450

#> Warning: non-integer x = 1.351560

#> Warning: non-integer x = 711.304000

#> Warning: non-integer x = 13.858130

#> Warning: non-integer x = 2.654355

#> Warning: non-integer x = 1510.784000

#> Warning: non-integer x = 65.571110

#> Warning: non-integer x = 6318.246000

#> Warning: non-integer x = 641.955600

#> Warning: non-integer x = 16.118900

#> Warning: non-integer x = 81.213000

#> Warning: non-integer x = 213.089000

#> Warning: non-integer x = 341.246400

#> Warning: non-integer x = 1668.030000

#> Warning: non-integer x = 299.616900

#> Warning: non-integer x = 0.000668

#> Warning: non-integer x = 20.262440

#> Warning: non-integer x = 340.883000

#> Warning: non-integer x = 1606.748000

#> Warning: non-integer x = 1353.561000

#> Warning: non-integer x = 567.751000

#> Warning: non-integer x = 3787.443000

#> Warning: non-integer x = 0.583718

#> Warning: non-integer x = 87.754000

#> Warning: non-integer x = 82.802100

#> Warning: non-integer x = 1037.524400

#> Warning: non-integer x = 0.698498

#> Warning: non-integer x = 38.514000

#> Warning: non-integer x = 234.856000

#> Warning: non-integer x = 0.857815

#> Warning: non-integer x = 34.385800

#> Warning: non-integer x = 39.990300

#> Warning: non-integer x = 190.953400

#> Warning: non-integer x = 545.017000

#> Warning: non-integer x = 0.419190

#> Warning: non-integer x = 224.375000

#> Warning: non-integer x = 1.578533

#> Warning: non-integer x = 33.842050

#> Warning: non-integer x = 878.895000

#> Warning: non-integer x = 3.379013

#> Warning: non-integer x = 10.428100

#> Warning: non-integer x = 0.263687

#> Warning: non-integer x = 696.973000

#> Warning: non-integer x = 174.165200

#> Warning: non-integer x = 933.904000

#> Warning: non-integer x = 20.656400

#> Warning: non-integer x = 2.682300

#> Warning: non-integer x = 242.526000

#> Warning: non-integer x = 1106.710000

#> Warning: non-integer x = 28031.200000

#> Warning: non-integer x = 5.807153

#> Warning: non-integer x = 2.825400

#> Warning: non-integer x = 3910.544000

#> Warning: non-integer x = 142.207000

#> Warning: non-integer x = 0.013718

#> Warning: non-integer x = 0.867050

#> Warning: non-integer x = 312.668400

#> Warning: non-integer x = 24.677400

#> Warning: non-integer x = 1605.263000

#> Warning: non-integer x = 22.221100

#> Warning: non-integer x = 8.211511

#> Warning: non-integer x = 3.760559

#> Warning: non-integer x = 6.947490

#> Warning: non-integer x = 59.002900

#> Warning: non-integer x = 12.576080

#> Warning: non-integer x = 186.360900

#> Warning: non-integer x = 5.371050

#> Warning: non-integer x = 68.888160

#> Warning: non-integer x = 1848.310000

#> Warning: non-integer x = 443.146000

#> Warning: non-integer x = 25.675600

#> Warning: non-integer x = 2.588400

#> Warning: non-integer x = 2332.682000

#> Warning: non-integer x = 15.658100

#> Warning: non-integer x = 3.839600

#> Warning: non-integer x = 159.878400

#> Warning: non-integer x = 76.351700

#> Warning: non-integer x = 5.202175

#> Warning: non-integer x = 29.761300

#> Warning: non-integer x = 9.806910

#> Warning: non-integer x = 25.242600

#> Warning: non-integer x = 1312.780000

#> Warning: non-integer x = 0.047590

#> Warning: non-integer x = 13.202860

#> Warning: non-integer x = 0.246652

#> Warning: non-integer x = 1051.528500

#> Warning: non-integer x = 482.882000

#> Warning: non-integer x = 7666.940000

#> Warning: non-integer x = 18.319290

#> Warning: non-integer x = 9.401030

#> Warning: non-integer x = 16.495300

#> Warning: non-integer x = 497.725800

#> Warning: non-integer x = 0.781422

#> Warning: non-integer x = 0.267752

#> Warning: non-integer x = 66.927300

#> Warning: non-integer x = 226.909000

#> Warning: non-integer x = 3971.870000

#> Warning: non-integer x = 2.690248

#> Warning: non-integer x = 5.940210

#> Warning: non-integer x = 304.073000

#> Warning: non-integer x = 158.713500

#> Warning: non-integer x = 245.523300

#> Warning: non-integer x = 153.159600

#> Warning: non-integer x = 410.987000

#> Warning: non-integer x = 1.075082

#> Warning: non-integer x = 1153.064000

#> Warning: non-integer x = 27.936150

#> Warning: non-integer x = 14.933600

#> Warning: non-integer x = 411.028200

#> Warning: non-integer x = 3.515270

#> Warning: non-integer x = 2801.491000

#> Warning: non-integer x = 16.957300

#> Warning: non-integer x = 9833.703000

#> Warning: non-integer x = 14.294800

#> Warning: non-integer x = 106.494000

#> Warning: non-integer x = 7.822265

#> Warning: non-integer x = 0.620290

#> Warning: non-integer x = 0.073214

#> Warning: non-integer x = 1.484990

#> Warning: non-integer x = 149.397600

#> Warning: non-integer x = 7.185640

#> Warning: non-integer x = 177.208900

#> Warning: non-integer x = 23.688830

#> Warning: non-integer x = 1545.995000

#> Warning: non-integer x = 36.091220

#> Warning: non-integer x = 30.270000

#> Warning: non-integer x = 30.087530

#> Warning: non-integer x = 127.055500

#> Warning: non-integer x = 773.707000

#> Warning: non-integer x = 1631.652000

#> Warning: non-integer x = 457.709000

#> Warning: non-integer x = 299.361000

#> Warning: non-integer x = 175.032000

#> Warning: non-integer x = 1828.992000

#> Warning: non-integer x = 148.534100

#> Warning: non-integer x = 54.499400

#> Warning: non-integer x = 9565.270000

#> Warning: non-integer x = 117.343600

#> Warning: non-integer x = 54.249100

#> Warning: non-integer x = 32.158600

#> Warning: non-integer x = 3.313233

#> Warning: non-integer x = 0.101149

#> Warning: non-integer x = 9.045322

#> Warning: non-integer x = 3169.265000

#> Warning: non-integer x = 0.877652

#> Warning: non-integer x = 58.683680

#> Warning: non-integer x = 4.873693

#> Warning: non-integer x = 65.061260

#> Warning: non-integer x = 13.167700

#> Warning: non-integer x = 247.348200

#> Warning: non-integer x = 0.395833

#> Warning: non-integer x = 31.245280

#> Warning: non-integer x = 52.084450

#> Warning: non-integer x = 29.989630

#> Warning: non-integer x = 163.721000

#> Warning: non-integer x = 0.934369

#> Warning: non-integer x = 2.701094

#> Warning: non-integer x = 6.119641

#> Warning: non-integer x = 8113.850000

#> Warning: non-integer x = 36.044700

#> Warning: non-integer x = 10.507200

#> Warning: non-integer x = 8.595202

#> Warning: non-integer x = 29.109630

#> Warning: non-integer x = 4.358050

#> Warning: non-integer x = 2.678040

#> Warning: non-integer x = 4.998301

#> Warning: non-integer x = 656.211000

#> Warning: non-integer x = 12.833260

#> Warning: non-integer x = 6.234430

#> Warning: non-integer x = 218.535000

#> Warning: non-integer x = 17.836000

#> Warning: non-integer x = 2998.402000

#> Warning: non-integer x = 1833.733000

#> Warning: non-integer x = 9.888340

#> Warning: non-integer x = 449.216000

#> Warning: non-integer x = 386.015300

#> Warning: non-integer x = 81.205850

#> Warning: non-integer x = 284.363000

#> Warning: non-integer x = 390.838800

#> Warning: non-integer x = 6.892853

#> Warning: non-integer x = 300.934700

#> Warning: non-integer x = 146.177900

#> Warning: non-integer x = 6.826700

#> Warning: non-integer x = 258.863000

#> Warning: non-integer x = 900.970000

#> Warning: non-integer x = 1.047274

#> Warning: non-integer x = 87.822600

#> Warning: non-integer x = 23.766490

#> Warning: non-integer x = 914.892000

#> Warning: non-integer x = 0.119381

#> Warning: non-integer x = 4.929111

#> Warning: non-integer x = 48.864200

#> Warning: non-integer x = 0.066107

#> Warning: non-integer x = 9.924740

#> Warning: non-integer x = 30.030900

#> Warning: non-integer x = 45.243660

#> Warning: non-integer x = 462.188000

#> Warning: non-integer x = 3.383006

#> Warning: non-integer x = 1.622300

#> Warning: non-integer x = 46.926600

#> Warning: non-integer x = 11.527560

#> Warning: non-integer x = 7.896361

#> Warning: non-integer x = 525.864900

#> Warning: non-integer x = 1.725361

#> Warning: non-integer x = 19.017800

#> Warning: non-integer x = 0.148183

#> Warning: non-integer x = 194.873000

#> Warning: non-integer x = 62.651400

#> Warning: non-integer x = 674.652000

#> Warning: non-integer x = 36.692300

#> Warning: non-integer x = 7.464600

#> Warning: non-integer x = 103.062410

#> Warning: non-integer x = 23.515300

#> Warning: non-integer x = 348420.600000

#> Warning: non-integer x = 4.616865

#> Warning: non-integer x = 10.309800

#> Warning: non-integer x = 764.860800

#> Warning: non-integer x = 204.514000

#> Warning: non-integer x = 0.524952

#> Warning: non-integer x = 0.397271

#> Warning: non-integer x = 26.813400

#> Warning: non-integer x = 14.755100

#> Warning: non-integer x = 636.416000

#> Warning: non-integer x = 19.511800

#> Warning: non-integer x = 38.067260

#> Warning: non-integer x = 4.996123

#> Warning: non-integer x = 110.706200

#> Warning: non-integer x = 983.807000

#> Warning: non-integer x = 72.686900

#> Warning: non-integer x = 3121.850000

#> Warning: non-integer x = 36.372600

#> Warning: non-integer x = 177.166000

#> Warning: non-integer x = 21061.100000

#> Warning: non-integer x = 3137.980000

#> Warning: non-integer x = 381.350200

#> Warning: non-integer x = 18.778300

#> Warning: non-integer x = 12841.400000

#> Warning: non-integer x = 1597.610000

#> Warning: non-integer x = 21.280600

#> Warning: non-integer x = 3391.971000

#> Warning: non-integer x = 1961.333000

#> Warning: non-integer x = 26.479200

#> Warning: non-integer x = 548.615000

#> Warning: non-integer x = 31.639300

#> Warning: non-integer x = 64.265300

#> Warning: non-integer x = 0.177100

#> Warning: non-integer x = 67.903000

#> Warning: non-integer x = 1.863400

#> Warning: non-integer x = 33477.700000

#> Warning: non-integer x = 3135.470000

#> Warning: non-integer x = 3491.370000

#> Warning: non-integer x = 252.677000

#> Warning: non-integer x = 210.273000

#> Warning: non-integer x = 265.872200

#> Warning: non-integer x = 1646.161000

#> Warning: non-integer x = 6.814500

#> Warning: non-integer x = 11.124300

#> Warning: non-integer x = 449.829000

#> Warning: non-integer x = 3286.670000

#> Warning: non-integer x = 54453.800000

#> Warning: non-integer x = 170.584000

#> Warning: non-integer x = 67.082400

#> Warning: non-integer x = 4317.370000

#> Warning: non-integer x = 441.775400

#> Warning: non-integer x = 2146.749000

#> Warning: non-integer x = 786.234900

#> Warning: non-integer x = 3273.952000

#> Warning: non-integer x = 41.739500

#> Warning: non-integer x = 15251.900000

#> Warning: non-integer x = 502.075200

#> Warning: non-integer x = 474.835000

#> Warning: non-integer x = 5454.174000

#> Warning: non-integer x = 74.792300

#> Warning: non-integer x = 19915.500000

#> Warning: non-integer x = 70.286700

#> Warning: non-integer x = 28765.700000

#> Warning: non-integer x = 103.332000

#> Warning: non-integer x = 883.234000

#> Warning: non-integer x = 193.058000

#> Warning: non-integer x = 178.965000

#> Warning: non-integer x = 6.243600

#> Warning: non-integer x = 45.130800

#> Warning: non-integer x = 2397.901000

#> Warning: non-integer x = 4.504500

#> Warning: non-integer x = 756.075100

#> Warning: non-integer x = 89.673100

#> Warning: non-integer x = 170978.500000

#> Warning: non-integer x = 245.957000

#> Warning: non-integer x = 1138.730000

#> Warning: non-integer x = 65.751400

#> Warning: non-integer x = 5422.550000

#> Warning: non-integer x = 10402.920000

#> Warning: non-integer x = 16047.240000

#> Warning: non-integer x = 4308.986000

#> Warning: non-integer x = 4926.890000

#> Warning: non-integer x = 323.576000

#> Warning: non-integer x = 22582.200000

#> Warning: non-integer x = 206.715000

#> Warning: non-integer x = 1075.480000

#> Warning: non-integer x = 5226.683000

#> Warning: non-integer x = 683.824000

#> Warning: non-integer x = 2324.322000

#> Warning: non-integer x = 767.443600

#> Warning: non-integer x = 1.300200

#> Warning: non-integer x = 0.699600

#> Warning: non-integer x = 49013.910000

#> Warning: non-integer x = 185.589000

#> Warning: non-integer x = 549.597400

#> Warning: non-integer x = 582.802000

#> Warning: non-integer x = 774.143000

#> Warning: non-integer x = 4.534200

#> Warning: non-integer x = 1218.217000

#> Warning: non-integer x = 70.885100

#> Warning: non-integer x = 610.847600

#> Warning: non-integer x = 482.796600

#> Warning: non-integer x = 2399.584000

#> Warning: non-integer x = 1726.571000

#> Warning: non-integer x = 34.804000

#> Warning: non-integer x = 244.837000

#> Warning: non-integer x = 16.947700

#> Warning: non-integer x = 26882.100000

#> Warning: non-integer x = 139.389000

#> Warning: non-integer x = 82.426300

#> Warning: non-integer x = 521.667000

#> Warning: non-integer x = 140.772500

#> Warning: non-integer x = 120.131000

#> Warning: non-integer x = 314.954000

#> Warning: non-integer x = 33.817300

#> Warning: non-integer x = 15887.300000

#> Warning: non-integer x = 146.477100

#> Warning: non-integer x = 78.940400

#> Warning: non-integer x = 3141.240000

#> Warning: non-integer x = 178.877000

#> Warning: non-integer x = 39005.700000

#> Warning: non-integer x = 3658.460000

#> Warning: non-integer x = 285.762400

#> Warning: non-integer x = 3238.150000

#> Warning: non-integer x = 4664.810000

#> Warning: non-integer x = 4255.152000

#> Warning: non-integer x = 1677.980000

#> Warning: non-integer x = 6311.624000

#> Warning: non-integer x = 0.754600

#> Warning: non-integer x = 139.042200

#> Warning: non-integer x = 4496.660000

#> Warning: non-integer x = 1495.813000

#> Warning: non-integer x = 1558.310000

#> Warning: non-integer x = 6693.170000

#> Warning: non-integer x = 17412.230000

#> Warning: non-integer x = 13.646600

#> Warning: non-integer x = 1558.560000

#> Warning: non-integer x = 206.149900

#> Warning: non-integer x = 27242.800000

#> Warning: non-integer x = 5.514300

#> Warning: non-integer x = 42.876900

#> Warning: non-integer x = 346.886100

#> Warning: non-integer x = 1.339800

#> Warning: non-integer x = 47.577200

#> Warning: non-integer x = 937.197000

#> Warning: non-integer x = 491.173100

#> Warning: non-integer x = 4921.160000

#> Warning: non-integer x = 7.945300

#> Warning: non-integer x = 5.809100

#> Warning: non-integer x = 1492.460000

#> Warning: non-integer x = 15.557300

#> Warning: non-integer x = 774.446200

#> Warning: non-integer x = 13800.900000

#> Warning: non-integer x = 336.273300

#> Warning: non-integer x = 178.695000

#> Warning: non-integer x = 4.120600

#> Warning: non-integer x = 181.184300

#> Warning: non-integer x = 393.505200

#> Warning: non-integer x = 9656.740000

#> Warning: non-integer x = 421.276000

#> Warning: non-integer x = 151.637200

#> Warning: non-integer x = 4095.200000

#> Warning: non-integer x = 447.580100

#> Warning: non-integer x = 446.712200

#> Warning: non-integer x = 16.325100

#> Warning: non-integer x = 1868.955000

#> Warning: non-integer x = 8215.174000

#> Warning: non-integer x = 20.156400

#> Warning: non-integer x = 14.248300

#> Warning: non-integer x = 886.811200

#> Warning: non-integer x = 330.186000

#> Warning: non-integer x = 7494.810000

#> Warning: non-integer x = 75.641500

#> Warning: non-integer x = 112.820400

#> Warning: non-integer x = 149.928900

#> Warning: non-integer x = 45.661660

#> Warning: non-integer x = 417.906500

#> Warning: non-integer x = 34.638600

#> Warning: non-integer x = 438.419300

#> Warning: non-integer x = 22.728800

#> Warning: non-integer x = 61.317630

#> Warning: non-integer x = 5156.740000

#> Warning: non-integer x = 3121.712000

#> Warning: non-integer x = 970.308900

#> Warning: non-integer x = 19.803600

#> Warning: non-integer x = 24286.130000

#> Warning: non-integer x = 76.497630

#> Warning: non-integer x = 5.479430

#> Warning: non-integer x = 191.959000

#> Warning: non-integer x = 456.058900

#> Warning: non-integer x = 38.385490

#> Warning: non-integer x = 184.709000

#> Warning: non-integer x = 16.826400

#> Warning: non-integer x = 12.301000

#> Warning: non-integer x = 1873.421000

#> Warning: non-integer x = 1.036980

#> Warning: non-integer x = 28.238210

#> Warning: non-integer x = 1.276957

#> Warning: non-integer x = 10532.100000

#> Warning: non-integer x = 8504.408000

#> Warning: non-integer x = 380.556000

#> Warning: non-integer x = 6659.664000

#> Warning: non-integer x = 135.273000

#> Warning: non-integer x = 46.373030

#> Warning: non-integer x = 54.554720

#> Warning: non-integer x = 270.920100

#> Warning: non-integer x = 0.178210

#> Warning: non-integer x = 8.373850

#> Warning: non-integer x = 68.385400

#> Warning: non-integer x = 2582.830000

#> Warning: non-integer x = 52441.600000

#> Warning: non-integer x = 10.419629

#> Warning: non-integer x = 11.583660

#> Warning: non-integer x = 6260.496000

#> Warning: non-integer x = 104.802170

#> Warning: non-integer x = 341.351000

#> Warning: non-integer x = 83.364710

#> Warning: non-integer x = 1188.044000

#> Warning: non-integer x = 4.009522

#> Warning: non-integer x = 21076.550000

#> Warning: non-integer x = 599.179900

#> Warning: non-integer x = 77.788300

#> Warning: non-integer x = 3552.692000

#> Warning: non-integer x = 10.581700

#> Warning: non-integer x = 48495.700000

#> Warning: non-integer x = 57.641400

#> Warning: non-integer x = 76.959960

#> Warning: non-integer x = 386.503000

#> Warning: non-integer x = 66.150200

#> Warning: non-integer x = 28.435800

#> Warning: non-integer x = 1.389091

#> Warning: non-integer x = 92.856500

#> Warning: non-integer x = 2823.315000

#> Warning: non-integer x = 14.621000

#> Warning: non-integer x = 58.784220

#> Warning: non-integer x = 41.450310

#> Warning: non-integer x = 5771.403000

#> Warning: non-integer x = 15.639910

#> Warning: non-integer x = 404.923000

#> Warning: non-integer x = 14.420700

#> Warning: non-integer x = 1888.800000

#> Warning: non-integer x = 643.326200

#> Warning: non-integer x = 5492.355000

#> Warning: non-integer x = 32863.710000

#> Warning: non-integer x = 872.211000

#> Warning: non-integer x = 385.015400

#> Warning: non-integer x = 17378.240000

#> Warning: non-integer x = 144.995400

#> Warning: non-integer x = 324.768000

#> Warning: non-integer x = 8130.177000

#> Warning: non-integer x = 426.481000

#> Warning: non-integer x = 435.874000

#> Warning: non-integer x = 6.400042

#> Warning: non-integer x = 10.212100

#> Warning: non-integer x = 0.360670

#> Warning: non-integer x = 17.099100

#> Warning: non-integer x = 3456.871000

#> Warning: non-integer x = 1.869956

#> Warning: non-integer x = 422.840000

#> Warning: non-integer x = 24.010400

#> Warning: non-integer x = 413.599000

#> Warning: non-integer x = 54.273450

#> Warning: non-integer x = 212.218600

#> Warning: non-integer x = 2.631046

#> Warning: non-integer x = 517.927000

#> Warning: non-integer x = 238.407400

#> Warning: non-integer x = 78.848800

#> Warning: non-integer x = 586.435300

#> Warning: non-integer x = 37.117200

#> Warning: non-integer x = 8.594560

#> Warning: non-integer x = 12.521630

#> Warning: non-integer x = 2354.350000

#> Warning: non-integer x = 37.168000

#> Warning: non-integer x = 8.664161

#> Warning: non-integer x = 7.218398

#> Warning: non-integer x = 23.017500

#> Warning: non-integer x = 46.427300

#> Warning: non-integer x = 91.353020

#> Warning: non-integer x = 21.237200

#> Warning: non-integer x = 1858.150000

#> Warning: non-integer x = 27.105540

#> Warning: non-integer x = 9.442257

#> Warning: non-integer x = 1698.081000

#> Warning: non-integer x = 6.375600

#> Warning: non-integer x = 29147.360000

#> Warning: non-integer x = 4257.682000

#> Warning: non-integer x = 13.294000

#> Warning: non-integer x = 738.644500

#> Warning: non-integer x = 990.513000

#> Warning: non-integer x = 220.760100

#> Warning: non-integer x = 110.090000

#> Warning: non-integer x = 492.944100

#> Warning: non-integer x = 0.117710

#> Warning: non-integer x = 9.852430

#> Warning: non-integer x = 4204.761000

#> Warning: non-integer x = 4015.275000

#> Warning: non-integer x = 34.821400

#> Warning: non-integer x = 1271.410000

#> Warning: non-integer x = 4187.854000

#> Warning: non-integer x = 7.341675

#> Warning: non-integer x = 314.180000

#> Warning: non-integer x = 384.253100

#> Warning: non-integer x = 4711.630000

#> Warning: non-integer x = 2.958824

#> Warning: non-integer x = 37.558900

#> Warning: non-integer x = 52.370500

#> Warning: non-integer x = 0.650220

#> Warning: non-integer x = 19.619380

#> Warning: non-integer x = 598.782000

#> Warning: non-integer x = 379.700000

#> Warning: non-integer x = 9614.891000

#> Warning: non-integer x = 11.984280

#> Warning: non-integer x = 33.527500

#> Warning: non-integer x = 165.566500

#> Warning: non-integer x = 6.253841

#> Warning: non-integer x = 282.938000

#> Warning: non-integer x = 1335.940000

#> Warning: non-integer x = 14.174800

#> Warning: non-integer x = 31.821300

#> Warning: non-integer x = 6.469342

#> Warning: non-integer x = 207.349000

#> Warning: non-integer x = 260.830900

#> Warning: non-integer x = 5134.410000

#> Warning: non-integer x = 174.391000

#> Warning: non-integer x = 79.139170

#> Warning: non-integer x = 684.453000

#> Warning: non-integer x = 81.710200

#> Warning: non-integer x = 64464.730000

#> Warning: non-integer x = 57.702810

#> Warning: non-integer x = 17.933520

#> Warning: non-integer x = 480.450300

#> Warning: non-integer x = 193.782600

#> Warning: non-integer x = 1.632983

#> Warning: non-integer x = 0.132196

#> Warning: non-integer x = 125.317500

#> Warning: non-integer x = 6.984930

#> Warning: non-integer x = 4427.093000

#> Warning: non-integer x = 34.057650

#> Warning: non-integer x = 112.876000

#> Warning: non-integer x = 59.081700