Tetrads

tetrads estimates gravity models by taking the ratio of the ratio of flows.

tetrads(dependent_variable, regressors, codes,
  reference_countries = c("JPN", "USA"), multiway = TRUE, data, ...)

Arguments

dependent_variable
name (type: character) of the dependent variable in the dataset data (e.g. trade flows).

This variable is logged and then used as the dependent variable in the estimation.

regressors
name (type: character) of the regressors to include in the model.

Include the distance variable in the dataset data containing a measure of distance between all pairs of bilateral partners and bilateral variables that should be taken as the independent variables in the estimation.

The distance is logged automatically when the function is executed.

Unilateral effects drop as the ratio of ratios is taken.

Write this argument as c(distance, contiguity, common curreny, ...).

codes
variable name (type: character) of the code of the country of origin and destination (e.g. ISO-3 codes from the variables iso_o and iso_d) in the example datasets).

The variables are grouped by using iso_o and iso_d to obtain estimates.

Write this argument as c(code origin, code destination).

reference_countries
reference exporting and importing country, default is set to c("JPN", "USA")

Write this argument as c(importing country, exporting country).

multiway
(type: logic) In case multiway = TRUE, the cluster.vcov function is used for estimation following Cameron et al. (2011) multi-way clustering of variance-covariance matrices.

The default value is set to TRUE.

data
name of the dataset to be used (type: character).

To estimate gravity equations you need a square dataset including bilateral flows defined by the argument dependent_variable, ISO codes or similar of type character (e.g. iso_o for the country of origin and iso_d for the destination country), a distance measure defined by the argument distance and other potential influences (e.g. contiguity and common currency) given as a vector in regressors are required.

All dummy variables should be of type numeric (0/1).

Make sure the ISO codes are of type "character".

If an independent variable is defined as a ratio, it should be logged.

The user should perform some data cleaning beforehand to remove observations that contain entries that can distort estimates.

When using panel data, a variable for the time may be included in the dataset. Note that the variable for the time dimension should be of type factor.

The time variable can be used as a single dependent variable or interaction term with other variables such as country identifiers by inserting it into regressors or as an optional parameter.

The function will remove zero flows and distances.

...
additional arguments to be passed to functions used by tetrads.

Value

The function returns the summary of the estimated gravity model as an lm-object.

Details

tetrads is an estimation method for gravity models developed by Head et al. (2010) .

The function tetrads utilizes the multiplicative form of the gravity equation. After choosing a reference exporter A and importer B one can eliminate importer and exporter fixed effects by taking the ratio of ratios.

Only those exporters trading with the reference importer and importers trading with the reference exporter will remain for the estimation. Therefore, reference countries should preferably be countries which trade with every other country in the dataset.

After restricting the data in this way, tetrads estimates the gravity equation in its additive form by OLS.

By taking the ratio of ratios, all monadic effects diminish, hence no unilateral variables such as GDP can be included as independent variables.

tetrads estimation can be used for both, cross-sectional as well as panel data. Nonetheless, the function is designed to be consistent with the Stata code for cross-sectional data provided on the website Gravity Equations: Workhorse, Toolkit, and Cookbook when choosing robust estimation.

The function tetrads was therefore tested for cross-sectional data.

If tetrads is used for panel data, the user may have to drop distance as an independent variable as time-invariant effects drop.

For applying tetrads to panel data see Head et al. (2010) .

References

For information on tetrads see

Cameron AC, Gelbach JB and Miller DL (2011). “Robust Inference With Multiway Clustering.” Journal of Business & Economic Statistics, 29(2), pp. 238--249. ISSN 07350015.

Head K, Mayer T and Ries J (2010). “The erosion of colonial trade linkages after independence.” Journal of International Economics, 81(1), pp. 1 - 14. ISSN 0022-1996, doi: 10.1016/j.jinteco.2010.01.002 .

For more information on gravity models, theoretical foundations and estimation methods in general see

Anderson JE (1979). “A Theoretical Foundation for the Gravity Equation.” The American Economic Review, 69(1), pp. 106--116. ISSN 00028282.

Anderson JE and van Wincoop E (2001). “Gravity with Gravitas: A Solution to the Border Puzzle.” Technical Report 8079, National Bureau of Economic Research. doi: 10.3386/w8079 .

Anderson JE (2010). “The Gravity Model.” Technical Report 16576, National Bureau of Economic Research. doi: 10.3386/w16576 .

Head K, Mayer T and Ries J (2010). “The erosion of colonial trade linkages after independence.” Journal of International Economics, 81(1), pp. 1 - 14. ISSN 0022-1996, doi: 10.1016/j.jinteco.2010.01.002 .

Head K and Mayer T (2014). “Chapter 3 - Gravity Equations: Workhorse,Toolkit, and Cookbook.” In Gopinath G, Helpman E and Rogoff K (eds.), Handbook of International Economics, volume 4 series Handbook of International Economics, pp. 131 - 195. Elsevier. doi: 10.1016/B978-0-444-54314-1.00003-3 .

Silva JMCS and Tenreyro S (2006). “The Log of Gravity.” The Review of Economics and Statistics, 88(4), pp. 641-658. doi: 10.1162/rest.88.4.641 .

and the citations therein.

See Gravity Equations: Workhorse, Toolkit, and Cookbook for gravity datasets and Stata code for estimating gravity models.

For estimating gravity equations using panel data see

Egger P and Pfaffermayr M (2003). “The proper panel econometric specification of the gravity equation: A three-way model with bilateral interaction effects.” Empirical Economics, 28(3), pp. 571--580. ISSN 1435-8921, doi: 10.1007/s001810200146 .

Gómez-Herrera E (2013). “Comparing alternative methods to estimate gravity models of bilateral trade.” Empirical Economics, 44(3), pp. 1087--1111. ISSN 1435-8921, doi: 10.1007/s00181-012-0576-2 .

and the references therein.

See also

Examples

# NOT RUN {
data(gravity_no_zeros)

tetrads(dependent_variable = "flow", regressors = c("distw", "rta"),
codes = c("iso_o", "iso_d"), reference_countries = c("JPN", "USA"),
multiway = TRUE, data = gravity_no_zeros)

tetrads(dependent_variable = "flow", regressors = c("distw", "rta", "comcur", "contig"),
codes = c("iso_o", "iso_d"), reference_countries = c("JPN", "USA"),
multiway = FALSE, data = gravity_no_zeros)
# }