Gamma Pseudo Maximum Likelihood (GPML)

gpml estimates gravity models in their multiplicative form via Gamma Pseudo Maximum Likelihood.

gpml(dependent_variable, distance, additional_regressors, robust = FALSE,
  data, ...)

Arguments

dependent_variable
(Type: character) name of the dependent variable. This variable is logged and then used as the dependent variable in the estimation.

distance
(Type: character) name of the distance variable that should be taken as the key independent variable in the estimation. The distance is logged automatically when the function is executed.

additional_regressors
(Type: character) names of the additional regressors to include in the model (e.g. a dummy variable to indicate contiguity). Unilateral metric variables such as GDPs can be added but those variables have to be logged first. Interaction terms can be added.

Write this argument as c(distance, contiguity, common currency, ...).

robust
(Type: logical) whether robust fitting should be used. By default this is set to FALSE.

data
(Type: data.frame) the dataset to be used.

...
Additional arguments to be passed to bvw.

Value

The function returns the summary of the estimated gravity model similar to a glm-object.

Details

gpml is an estimation method for gravity models belonging to generalized linear models. It is described in Silva and Tenreyro (2006) and the model is estimated via glm2 using the gamma distribution and a log-link.

For similar functions, utilizing the multiplicative form via the log-link, but different distributions, see ppml, nls, and nbpml.

gpml estimation can be used for both, cross-sectional as well as panel data, but its up to the user to ensure that the functions can be applied to panel data.

Depending on the panel dataset and the variables - specifically the type of fixed effects - included in the model, it may easily occur that the model is not computable.

Also, note that by including bilateral fixed effects such as country-pair effects, the coefficients of time-invariant observables such as distance can no longer be estimated.

Depending on the specific model, the code of the respective function might have to be changed in order to exclude the distance variable from the estimation.

At the very least, the user should take special care with respect to the meaning of the estimated coefficients and variances as well as the decision about which effects to include in the estimation.

When using panel data, the parameter and variance estimation of the models may have to be changed accordingly.

For a comprehensive overview of gravity models for panel data see Egger and Pfaffermayr (2003) , Gómez-Herrera (2013) and Head et al. (2010) as well as the references therein.

References

Egger P and Pfaffermayr M (2003). “The proper panel econometric specification of the gravity equation: A three-way model with bilateral interaction effects.” Empirical Economics, 28(3), pp. 571--580. ISSN 1435-8921, doi: 10.1007/s001810200146 .

Gómez-Herrera E (2013). “Comparing alternative methods to estimate gravity models of bilateral trade.” Empirical Economics, 44(3), pp. 1087--1111. ISSN 1435-8921, doi: 10.1007/s00181-012-0576-2 .

Head K, Mayer T and Ries J (2010). “The erosion of colonial trade linkages after independence.” Journal of International Economics, 81(1), pp. 1 - 14. ISSN 0022-1996, doi: 10.1016/j.jinteco.2010.01.002 .

Silva JMCS and Tenreyro S (2006). “The Log of Gravity.” The Review of Economics and Statistics, 88(4), pp. 641-658. doi: 10.1162/rest.88.4.641 .

See also

Examples

# Example for CRAN checks:
# Executable in < 5 sec
library(dplyr)
data("gravity_no_zeros")

# Choose 5 countries for testing
countries_chosen <- c("AUS", "CHN", "GBR", "BRA", "CAN")
grav_small <- filter(gravity_no_zeros, iso_o %in% countries_chosen)

grav_small <- grav_small %>%
  mutate(
    lgdp_o = log(gdp_o),
    lgdp_d = log(gdp_d)
  )

fit <- gpml(
  dependent_variable = "flow",
  distance = "distw",
  additional_regressors = c("rta", "iso_o", "iso_d"),
  robust = FALSE,
  data = grav_small
)