Gamma Pseudo Maximum Likelihood (GPML)

gpml estimates gravity models in their multiplicative form via Gamma Pseudo Maximum Likelihood.

gpml(dependent_variable, regressors, robust = TRUE, data, ...)

Arguments

dependent_variable
name (type: character) of the dependent variable in the dataset data (e.g. trade flows).

regressors
name (type: character) of the regressors to include in the model.

Include the distance variable in the dataset data containing a measure of distance between all pairs of bilateral partners and bilateral variables that should be taken as the independent variables in the estimation.

The distance is logged automatically when the function is executed.

Unilateral metric variables such as GDPs can be added but those variables have to be logged first.

Interaction terms can be added.

Write this argument as c(distance, contiguity, common curreny, ...).

robust
robust (type: logical) determines whether a robust variance-covariance matrix should be used. By default is set to TRUE.

data
name of the dataset to be used (type: character).

To estimate gravity equations you need a square dataset including bilateral flows defined by the argument dependent_variable, ISO codes or similar of type character (e.g. iso_o for the country of origin and iso_d for the destination country), a distance measure defined by the argument distance and other potential influences (e.g. contiguity and common currency) given as a vector in regressors are required.

All dummy variables should be of type numeric (0/1).

Make sure the ISO codes are of type "character".

If an independent variable is defined as a ratio, it should be logged.

The user should perform some data cleaning beforehand to remove observations that contain entries that can distort estimates.

When using panel data, a variable for the time may be included in the dataset. Note that the variable for the time dimension should be of type factor.

The function will remove zero flows and distances.

...
additional arguments to be passed to functions used by gpml.

Value

The function returns the summary of the estimated gravity model similar to a glm-object.

Details

gpml is an estimation method for gravity models belonging to generalized linear models.

It is estimated via glm2 using the gamma distribution and a log-link.

For similar functions, utilizing the multiplicative form via the log-link, but different distributions, see ppml, nls, and nbpml.

gpml estimation can be used for both, cross-sectional as well as panel data.

It is up to the user to ensure that the functions can be applied to panel data.

Depending on the panel dataset and the variables - specifically the type of fixed effects - included in the model, it may easily occur that the model is not computable.

Also, note that by including bilateral fixed effects such as country-pair effects, the coefficients of time-invariant observables such as distance can no longer be estimated.

Depending on the specific model, the code of the respective function may has to be changed in order to exclude the distance variable from the estimation.

At the very least, the user should take special care with respect to the meaning of the estimated coefficients and variances as well as the decision about which effects to include in the estimation.

When using panel data, the parameter and variance estimation of the models may have to be changed accordingly.

For a comprehensive overview of gravity models for panel data see Egger and Pfaffermayr (2003) , Gómez-Herrera (2013) and Head et al. (2010) as well as the references therein.

References

For more information on the estimation of gravity equations via Gamma Pseudo maximum Likelihood see

Santos-Silva, J. M. C. and Tenreyro, S. (2006) <DOI:10.1162/rest.88.4.641>

For more information on gravity models, theoretical foundations and estimation methods in general see

Anderson JE (1979). “A Theoretical Foundation for the Gravity Equation.” The American Economic Review, 69(1), pp. 106--116. ISSN 00028282.

Anderson JE and van Wincoop E (2001). “Gravity with Gravitas: A Solution to the Border Puzzle.” Technical Report 8079, National Bureau of Economic Research. doi: 10.3386/w8079 .

Anderson JE (2010). “The Gravity Model.” Technical Report 16576, National Bureau of Economic Research. doi: 10.3386/w16576 .

Baier SL and Bergstrand JH (2009). “Bonus vetus OLS: A simple method for approximating international trade-cost effects using the gravity equation.” Journal of International Economics, 77(1), pp. 77 - 85. ISSN 0022-1996, doi: 10.1016/j.jinteco.2008.10.004 .

Baier SL and Bergstrand JH (2010). “The Gravity Model in International Trade: Advances and Applications.” In van Bergeijk PAG and Brakman S (eds.), chapter 4. Cambridge University Press. doi: 10.1017/CBO9780511762109 .

Head K, Mayer T and Ries J (2010). “The erosion of colonial trade linkages after independence.” Journal of International Economics, 81(1), pp. 1 - 14. ISSN 0022-1996, doi: 10.1016/j.jinteco.2010.01.002 .

Head K and Mayer T (2014). “Chapter 3 - Gravity Equations: Workhorse,Toolkit, and Cookbook.” In Gopinath G, Helpman E and Rogoff K (eds.), Handbook of International Economics, volume 4 series Handbook of International Economics, pp. 131 - 195. Elsevier. doi: 10.1016/B978-0-444-54314-1.00003-3 .

Silva JMCS and Tenreyro S (2006). “The Log of Gravity.” The Review of Economics and Statistics, 88(4), pp. 641-658. doi: 10.1162/rest.88.4.641 .

and the citations therein.

See Gravity Equations: Workhorse, Toolkit, and Cookbook for gravity datasets and Stata code for estimating gravity models.

For estimating gravity equations using panel data see

Egger P and Pfaffermayr M (2003). “The proper panel econometric specification of the gravity equation: A three-way model with bilateral interaction effects.” Empirical Economics, 28(3), pp. 571--580. ISSN 1435-8921, doi: 10.1007/s001810200146 .

Gómez-Herrera E (2013). “Comparing alternative methods to estimate gravity models of bilateral trade.” Empirical Economics, 44(3), pp. 1087--1111. ISSN 1435-8921, doi: 10.1007/s00181-012-0576-2 .

and the references therein.

See also

Examples

# NOT RUN {
data(gravity_no_zeros)

gravity_no_zeros <- gravity_no_zeros %>%
   mutate(
     lgdp_o = log(gdp_o),
     lgdp_d = log(gdp_d)
   )

gpml(dependent_variable = "flow", regressors = c("distw", "rta", "lgdp_o", "lgdp_d"),
robust = TRUE, data = gravity_no_zeros)
# }