Gravity dataset without zero trade flows

An edited version of the full gravity dataset: The "square" gravity dataset for all possible pairs of Countries worldwide, 1948-2006, which is used in the article: Head, K., T. Mayer and J. Ries, 2010, "The erosion of colonial linkages after independence". Journal of International Economics, 81(1):1-14 (lead article). Dataset gravity_no_zeros corresponds to the dataset without zero trade flows, gravity_zeros, on the other hand, includes zero trade flows.

data("gravity_zeros")

Format

A data frame with 22588 observations on the following 10 variables.

iso_o

ISO-Code of country of origin

iso_d

ISO-Code of country of destination

distw

weighted distance

gdp_o

GDP of country of origin

gdp_d

GDP of country of destination

rta

regional trade agreement

flow

trade flow

contig

contiguity

comlang_off

common official language

comcur

common currency

Source

http://econ.sciences-po.fr/node/131 https://sites.google.com/site/hiegravity/data-sources

References

Head, K. and Mayer, T. (2014). Chapter 3 - gravity equations: Workhorse,toolkit, and cookbook. In Gita Gopinath, E. H. and Rogoff, K., editors, Handbook of International Economics, volume 4 of Handbook of International Economics, pages 131-195. Elsevier. (Gravity Equations: Workhorse, Toolkit, and Cookbook)

Head, K., T. Mayer and J. Ries, 2010, "The erosion od colonial linkages after independence". Journal of International Economics, 81(1):1-14 (lead article).

Examples

  data(gravity_zeros)
  str(gravity_zeros)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	22588 obs. of  10 variables:
#>  $ iso_o      : chr  "AFG" "AFG" "AFG" "AFG" ...
#>  $ iso_d      : chr  "ARG" "AUS" "AUT" "AZE" ...
#>  $ distw      : num  15341 11086 4567 1822 5849 ...
#>  $ gdp_o      : num  0.0084 0.0084 0.0084 0.0084 0.0084 ...
#>  $ gdp_d      : num  0.214058 0.768178 0.322444 0.020122 0.000807 ...
#>  $ rta        : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ flow       : num  0.061 0.4055 0.1649 0.0016 0 ...
#>  $ contig     : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ comlang_off: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ comcur     : num  0 0 0 0 0 0 0 0 0 0 ...