Tetrads

tetrads estimates gravity models by taking the ratio of the ratio of flows.

tetrads(dependent_variable, distance, additional_regressors, code_origin,
  code_destination, filter_origin = NULL, filter_destination = NULL,
  multiway = TRUE, data, ...)

Arguments

dependent_variable
(Type: character) name of the dependent variable. This variable is logged and then used as the dependent variable in the estimation.

distance
(Type: character) name of the distance variable that should be taken as the key independent variable in the estimation. The distance is logged automatically when the function is executed.

additional_regressors
(Type: character) names of the additional regressors to include in the model (e.g. a dummy variable to indicate contiguity). Unilateral metric variables such as GDP should be inserted via the arguments income_origin and income_destination.

Write this argument as c(contiguity, common currency, ...). By default this is set to NULL.

code_origin
(Type: character) country of origin variable (e.g. ISO-3 country codes). The variables are grouped using this parameter.

code_destination
(Type: character) country of destination variable (e.g. country ISO-3 codes). The variables are grouped using this parameter.

filter_origin
(Type: character) Reference exporting country.

filter_destination
(Type: character) Reference importing country.

multiway
(Type: logical) in case multiway = TRUE, the cluster.vcov function is used for estimation following Cameron et al. (2011) multi-way clustering of variance-covariance matrices. The default value is set to TRUE.

data
(Type: data.frame) the dataset to be used.

...
Additional arguments to be passed to the function.

Value

The function returns the summary of the estimated gravity model as an lm-object.

Details

tetrads is an estimation method for gravity models developed by Head et al. (2010) .

The function tetrads utilizes the multiplicative form of the gravity equation. After choosing a reference exporter A and importer B one can eliminate importer and exporter fixed effects by taking the ratio of ratios.

Only those exporters trading with the reference importer and importers trading with the reference exporter will remain for the estimation. Therefore, reference countries should preferably be countries which trade with every other country in the dataset.

After restricting the data in this way, tetrads estimates the gravity equation in its additive form by OLS.

By taking the ratio of ratios, all monadic effects diminish, hence no unilateral variables such as GDP can be included as independent variables.

tetrads estimation can be used for both, cross-sectional as well as panel data. Nonetheless, the function is designed to be consistent with the Stata code for cross-sectional data provided on the website Gravity Equations: Workhorse, Toolkit, and Cookbook when choosing robust estimation.

The function tetrads was therefore tested for cross-sectional data.

If tetrads is used for panel data, the user may have to drop distance as an independent variable as time-invariant effects drop.

For applying tetrads to panel data see Head et al. (2010) .

References

For more information on gravity models, theoretical foundations and estimation methods in general see

Anderson JE (1979). “A Theoretical Foundation for the Gravity Equation.” The American Economic Review, 69(1), 106--116. ISSN 00028282.

Anderson JE, van Wincoop E (2001). “Gravity with Gravitas: A Solution to the Border Puzzle.” Technical Report 8079, National Bureau of Economic Research. doi: 10.3386/w8079 .

Anderson JE (2010). “The Gravity Model.” Technical Report 16576, National Bureau of Economic Research. doi: 10.3386/w16576 .

Baier SL, Bergstrand JH (2009). “Bonus vetus OLS: A simple method for approximating international trade-cost effects using the gravity equation.” Journal of International Economics, 77(1), 77 - 85. ISSN 0022-1996, doi: 10.1016/j.jinteco.2008.10.004 .

Baier SL, Bergstrand JH (2010). “The Gravity Model in International Trade: Advances and Applications.” In van Bergeijk PAG, Brakman S (eds.), chapter 4. Cambridge University Press. doi: 10.1017/CBO9780511762109 .

Feenstra RC (2002). “Border effects and the gravity equation: consistent methods for estimation.” Scottish Journal of Political Economy, 49(5), 491--506.

Head K, Mayer T, Ries J (2010). “The erosion of colonial trade linkages after independence.” Journal of International Economics, 81(1), 1 - 14. ISSN 0022-1996, doi: 10.1016/j.jinteco.2010.01.002 .

Head K, Mayer T (2014). “Chapter 3 - Gravity Equations: Workhorse,Toolkit, and Cookbook.” In Gopinath G, Helpman E, Rogoff K (eds.), Handbook of International Economics, volume 4 of Handbook of International Economics, 131 - 195. Elsevier. doi: 10.1016/B978-0-444-54314-1.00003-3 .

Silva JMCS, Tenreyro S (2006). “The Log of Gravity.” The Review of Economics and Statistics, 88(4), 641-658. doi: 10.1162/rest.88.4.641 .

and the citations therein.

See Gravity Equations: Workhorse, Toolkit, and Cookbook for gravity datasets and Stata code for estimating gravity models.

For estimating gravity equations using panel data see

Egger P, Pfaffermayr M (2003). “The proper panel econometric specification of the gravity equation: A three-way model with bilateral interaction effects.” Empirical Economics, 28(3), 571--580. ISSN 1435-8921, doi: 10.1007/s001810200146 .

Gómez-Herrera E (2013). “Comparing alternative methods to estimate gravity models of bilateral trade.” Empirical Economics, 44(3), 1087--1111. ISSN 1435-8921, doi: 10.1007/s00181-012-0576-2 .

and the references therein.

See also

Examples

# Example for CRAN checks:
# Executable in < 5 sec
library(dplyr)
data("gravity_no_zeros")

# Choose 5 countries for testing
countries_chosen <- c("AUS", "CHN", "GBR", "BRA", "CAN")
grav_small <- filter(gravity_no_zeros, iso_o %in% countries_chosen)

fit <- tetrads(
  dependent_variable = "flow",
  distance = "distw",
  additional_regressors = "rta",
  code_origin = "iso_o",
  code_destination = "iso_d",
  filter_origin = countries_chosen[1],
  filter_destination = countries_chosen[2],
  multiway = FALSE,
  data = grav_small
)