Left-censored Tobit model with known threshold

tobit estimates gravity models in their additive form by conducting a left-censored regression, which, after adding the constant 1 to the dependent variable, utilizes log(1)=0 as the censoring value.

tobit(dependent_variable, regressors, added_constant = 1, data, ...)

Arguments

dependent_variable
name (type: character) of the dependent variable in the dataset data, e.g. trade flows. The number 1 is added and the transformed variable is logged and taken as the dependent variable in the tobit estimation with lower bound equal to 0 as log(1)=0 represents the smallest flows in the transformed variable.

regressors
name (type: character) of the distance variable in the dataset data containing a measure of distance between all pairs of bilateral partners. It is logged automatically when the function is executed.

added_constant
scalar (type: numeric); represents the constant to be added to the dependent variable. The default value is 1. The minimum of log(y+added_constant) is taken as the left boundary in the Tobit model. In the often used case of added_constant=1, the dependent variable is left-censored at value 0 as log(1)=0.

data
name of the dataset to be used (type: character). To estimate gravity equations, a square gravity dataset including bilateral flows defined by the argument y, ISO-codes of type character (called iso_o for the country of origin and iso_d for the destination country), a distance measure defined by the argument dist and other potential influences given as a vector in x are required. All dummy variables should be of type numeric (0/1). Missing trade flows as well as incomplete rows should be excluded from the dataset. Zero trade flows are allowed.

...
additional arguments to be passed to tobit.

Value

The function returns the summary of the estimated gravity model as a censReg-object.

Details

tobit represents the left-censored tobit (Tobin, 1958) approach utilizing a known censoring threshold which is often used when several gravity models are compared. When taking the log of the gravity equation flows equal to zero constitute a problem as their log is not defined. Therefore, in the execution of the function the number 1 is added to all flows and the log(flows+1) is taken as the dependent variable. The tobit estimation is conducted using the censReg function and setting the lower bound equal to 0 as log(1)=0 represents the smallest flows in the transformed variable. A tobit regression represents a combination of a binary and a linear regression. This procedure has to be taken into consideration when interpreting the estimated coefficients. The marginal effects of an explanatory variable on the expected value of the dependent variable equals the product of both the probability of the latent variable exceeding the threshold and the marginal effect of the explanatory variable of the expected value of the latent variable.

To execute the function a square gravity dataset with all pairs of countries, ISO-codes for the country of origin and destination, a measure of distance between the bilateral partners as well as all information that should be considered as dependent an independent variables is needed. Missing bilateral flows as well as incomplete rows should be excluded from the dataset. Zero trade flows are allowed.

Up to now, the function is designed for cross-sectional data, but can be easily extended to panel data using the censReg function. A robust estimations is not implemented to the present as the censReg function is not compatible with the vcovHC function.

For more elaborate tobit functions, see et_tobit for the Eaton and Tamura (1994) threshold tobit model where instead of simply adding number 1 to the data the threshold is estimated or ek_tobit for the Eaton and Kortum (2001) tobit model where each zero trade volume is assigned a country specific interval with the upper bound equal to the minimum positive trade level of the respective importing country.

References

For more information on gravity models, theoretical foundations and estimation methods in general see

Anderson, J. E. (1979) <DOI:10.12691/wjssh-2-2-5>

Anderson, J. E. (2010) <DOI:10.3386/w16576>

Anderson, J. E. and van Wincoop, E. (2003) <DOI:10.3386/w8079>

Baier, S. L. and Bergstrand, J. H. (2009) <DOI:10.1016/j.jinteco.2008.10.004>

Baier, S. L. and Bergstrand, J. H. (2010) in Van Bergeijk, P. A., & Brakman, S. (Eds.) (2010) chapter 4 <DOI:10.1111/j.1467-9396.2011.01000.x>

Head, K., Mayer, T., & Ries, J. (2010) <DOI:10.1016/j.jinteco.2010.01.002>

Head, K. and Mayer, T. (2014) <DOI:10.1016/B978-0-444-54314-1.00003-3>

Santos-Silva, J. M. C. and Tenreyro, S. (2006) <DOI:10.1162/rest.88.4.641>

and the citations therein.

Especially for tobit models see

Tobin, J. (1958) <DOI:10.2307/1907382>

Eaton, J., & Tamura, A. (1994) <DOI:10.3386/w4758>

Eaton, J., & Kortum, S. (2001) <DOI:10.3386/w8070>.

See Gravity Equations: Workhorse, Toolkit, and Cookbook for gravity datasets and Stata code for estimating gravity models.

See also

Examples

# NOT RUN {
# Example for data with zero trade flows
data(gravity_zeros)

gravity_zeros <- gravity_zeros %>%
    mutate(
        lgdp_o = log(gdp_o),
        lgdp_d = log(gdp_d)
    )

tobit(dependent_variable = "flow", regressors = c("distw", "rta", "lgdp_o", "lgdp_d"),
added_constant = 1, data = gravity_zeros)
# }